padariaapp/requirements.txt
Django>=5.0.6,<5.1
psycopg2-binary==2.9.9
pillow>=10.3,<10.4
djangorestframework>=3.15,<3.16